requirements.txt
pytorch>=1.2
tensorboard>=1.14
skimage==0.16.2
cv2
hdf5
tqdm
